Extract Syntax enum
