+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod formatting;
 pub mod language;
 pub mod parsing;
 pub mod regex;
+pub mod rendering;
